template: split into main.rs and lib.rs
Split functionality between library crate and an executable crate that
consumes the library.
